requirements.txt
langchain-openai>=0.1.0
langchain-core>=0.2.0
langgraph>=0.2.0
typing-extensions>=4.0.0
python-dotenv>=0.19.0
